Lisätty chopconf write read + bugi korjattu spi write-funktiosta

+++ STM32F410/Inc/tmc2130.h
@@ -22,7 +22,7 @@ typedef struct {
 
 // prototyypit
 static uint32_t read_register(tmc2130 *tmc, uint8_t address);
-static void write_register(tmc2130 *tmc, uint8_t address, uint8_t value);
+static void write_register(tmc2130 *tmc, uint8_t address, uint32_t value);
 uint8_t tmc2130_init(tmc2130 *tmc, SPI_HandleTypeDef *spi, 
     GPIO_TypeDef *enable_port,
     GPIO_TypeDef *direction_port,
